+++ lib/broadside.rb
@@ -1,9 +1,9 @@
 require 'broadside/error'
 require 'broadside/utils'
 require 'broadside/configuration'
-require 'broadside/configuration/struct'
-require 'broadside/configuration/aws'
-require 'broadside/configuration/base'
+require 'broadside/configuration/config_struct'
+require 'broadside/configuration/aws_config'
+require 'broadside/configuration/base_config'
 require 'broadside/configuration/deploy_config'
 require 'broadside/configuration/ecs_config'
 require 'broadside/deploy'
