+++ lib/broadside/configuration/aws_config.rb
