+++ lib/broadside/configuration/base_config.rb
