+++ lib/broadside/configuration/config_struct.rb
